==> Human_NSCLC_dataset/1.Recluster.ipynb <==
import numpy as np
import pandas as pd
import scanpy as sc

sc.settings.verbosity = 3             # verbosity: errors (0), warnings (1), info (2), hints (3)
sc.logging.print_header()
sc.settings.set_figure_params(dpi=100, facecolor='white')

adata = sc.read_h5ad('/Users/bkim6/Desktop/BJ/NSCLC_atlas/BJ/adata_tumor.h5ad')

sc.pl.umap(adata, color=['leiden_1.00'], legend_loc='on data')

adata_15 = adata[adata.obs['leiden_1.00'] == '15']

#subcluster of cluster 15

sc.tl.leiden(adata_15, resolution=0.08, key_added='leiden_1.00')

sc.pl.umap(adata_15, color=['leiden_1.00'], legend_loc='right margin')

new_cluster_names = ['33', '15', '34']
adata_15.rename_categories('leiden_1.00', new_cluster_names)

sc.pl.umap(adata_15, color=['ASCL1', 'SYP', 'INSM1', 'CALCA', 'BEX1', 
                           'CHGA', 'SYP'])

adata_subset = adata[adata.obs['leiden_1.00'].isin(['0',
    '1',
'2',
'3',
'4',
'5',
'6',
'7',
'8',
'9',
'10',
'11',
'12',
'13',
'14',
'16',
'17',
'18',
'19',
'20',
'21',
'22',
'23',
'24',
'25',
'26',
'27',
'28',
'29',
'30',
'31',
'32'])]

#merge data

adata_re2 = adata_subset.concatenate(adata_15, index_unique=None, join='outer')

for col in adata_re2.obs.columns:
    if pd.api.types.is_categorical_dtype(adata_re2.obs[col]):
        adata_re2.obs[col]=adata_re2.obs[col].cat.add_categories("selfdefinedNA").fillna('selfdefinedNA')

sc.pl.umap(adata_re2, color=['leiden_1.00'], legend_loc='on data')

adata_re2.obs = adata_re2.obs[['sample', 'patient', 'tissue', 'condition', 'dataset', 'cell_type', 'leiden_1.00']]

adata_re2.write('/Users/bkim6/Desktop/BJ/NSCLC_atlas/BJ/adata_tumor_recluster.h5ad', compression='gzip')
